fix: query 폴더 추가
