+++ src/query/credits.ts
@@ -1,0 +1,8 @@
+import { useQuery } from "@tanstack/react-query";
+import { getCredits } from "../apis/axiosClient";
+
+export const useQueryGetCredits = (id: string) =>
+  useQuery({
+    queryKey: ["credits", id],
+    queryFn: () => getCredits(id),
+  });
