+++ src/query/query.ts
@@ -1,16 +1,13 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import {
   getMainList,
   getSearchList,
   getDetailList,
   getGenreList,
-  postFavorite,
-  getCredits,
-  getReviewList,
   getUpcomingMovies,
   getAiringTodayTVShows,
   getTvList,
-} from "./axiosClient";
+} from "../apis/axiosClient";
 
 export const useQueryGetMainList = () =>
   useQuery({
@@ -43,30 +40,6 @@ export const useQueryGetGenreList = () =>
   useQuery({
     queryKey: ["genre-list"],
     queryFn: () => getGenreList(),
-  });
-
-export const useMutationPostFavorite = (movieId: number) =>
-  useMutation({
-    mutationKey: ["post-favorite", movieId],
-    mutationFn: () => postFavorite(movieId),
-    onSuccess: () => {
-      alert("Successfully added to favorites!");
-    },
-    onError: (error) => {
-      console.error("Error posting favorite:", error);
-    },
-  });
-
-export const useQueryGetCredits = (id: string) =>
-  useQuery({
-    queryKey: ["credits", id],
-    queryFn: () => getCredits(id),
-  });
-
-export const useQueryGetReviewList = (id: string) =>
-  useQuery({
-    queryKey: ["review-list", id],
-    queryFn: () => getReviewList(id),
   });
 
 export const useQueryGetAiringTodayTVShows = () =>
